ftdetect: support user-overrides more thoroughly
Using
  let g:var = get(g:, 'var', default)
is an idiom for allowing a user to set the variable themselves but have
a default intact. In the case of lists/dictionaries, I use an empty
default and then extend() with the actual default: this essentially
merges the default with the user value. In the list case, there is no
way to not get the defaults; here, this should not be harmful. In the
dictionary case, I specify the 'keep' argument so that user values
override the defaults in case of a collision on keys.

Note that it is probably silly to override
g:racket_hash_lang_modifiers_regex or g:racket_hash_lang_regexp, but I
allow it anyway.

Related: https://github.com/wlangstroth/vim-racket/issues/62