fix for/fold indent for multi-line bodies
When the for/fold body is multi-line (unlike previous test cases with a
one-line cons) the IsSecondForm predicate incorrectly reported the start
of the body as being part of the second form.

I think it has to do with the loop over lnum <= vlnum, and that using <
might be a better way to go. For now this seems to work, though, so
stick with it.